Move unused ReflectionStrategies to benchmark project
